implemented support for downloading different packs
implemented using os.Args and generating a "config identifier" based on the repo name + filename + destination filepath.

it is using the 3rd to last directory in the filepath as that gives Prism Launcher users more compatibility for using this script in a lot of instances. That 3rd to last directory is the instance name.